requirements.txt
supabase>=2.0.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
resend>=0.6.0
feedparser>=6.0.0
arxiv>=2.0.0
httpx>=0.24.0
